Add vim to all images (#19)
When troubleshooting CI related issues, it is **really** useful to have a text editor handy.

We could select something smaller than `vim` (which is 55MB in Debian/Ubuntu 😢) but with a similar user interface, for example `nvi` but then RHEL does not have it, or we could select something smaller than `vim` but wit a non-standard UI like `joe` or `nano`, but if you are troubleshooting build issues you probably also know `vi` or `vim` better than the other two.