Switch to RHEL UBI image, add 9.4 (#29)
For creating the RHEL images, the RHEL `s2i-base` image was originally used as base, under the assumption it contained vital tools needed to create derived builder images, such as ours. However, looking more closely,
`s2i-base` is based on `s2i-core`, which in turn is based on `ubi`. Rebuilding the image using the latter image works perfectly fine, making it thus even smaller than before.

As an added bonus, there is no 9.4 tag for the `s2i-*` images, but exists for `ubi`. Hence, this change also adds an image for RHEL 9.4 using the lowest GCC we support - 12 - and the default Clang that is available via the package manager.