Add compiler name and version to package ID (#43)
We are running into the issue that Conan dependencies do not distinguish
between compilers. This has resulted in situations where a dependency
such as `b2` was built using GCC 15 and uploaded to the Conan cache, and
then pulled down for a build that used GCC 13, resulting in
compilation/linker failures.

This change adds the compiler name and version to the
`tools.info.package_id:confs` list. After building with this change the
entry in the cache has been updated to:

```
$> conan list 'b2/*:*'
Found 1 pkg/version recipes matching b2/* in local cache
Local Cache
  b2
    b2/5.3.3
      revisions
        107c15377719889654eb9a162a673975 (2025-08-05 16:41:05 UTC)
          packages
            ed959ecbcaea3dddbdbb9aa845785f37c5add3d2
              info
                settings
                  arch: armv8
                  os: Linux
                conf
                  user.package:compiler: "gcc-12"
                  user.package:distro: "debian-bookworm"
```

In addition to the above change, the README has been improved by adding
a `conan add remote` step, as well as updating the `conan install` step
to reflect what our actual CI pipelines run. Also a troubleshooting step
has been added for when the `docker run` command fails.