Add more perl packages to RHEL (#45)
The following Perl modules are also needed on RHEL. In this case I built
the image locally and pushed it to the repo, and then reran the failing
runs until all needed modules had been added.

Note that `perl` can also be installed in RHEL as a standalone package,
but it didn't install any or all of the needed modules.